requirements.txt
flask == 3.1.0
flask_caching == 2.3.1
redis == 6.0.0
Flask-WTF  ==1.2.2
pymongo == 4.12.1
PyYAML   == 6.0.2
Jinja2     ==3.1.6
gunicorn 
WTForms    ==3.2.1
Flask-Cors ==5.0.1
python-dotenv ==1.1.0